feat(trigger): add `MemTriggerStore` that can be used by Listeners to track triggers
Found myself re-writing the `runtimeListener.Listen` function for `cronListener.Listen`, and figured we could share code & test more thoroughly. In the future, if we feel we need different ways to store/track the triggers ephemerally, we can create an interface & allow options for the kind of mem store we need.