test(wftest): add `wftest.TestCase` that can be loaded from a dir
Because there is such a close relationship between a workflow and a dataset, we usually want both a workflow and dataset in a test case. We also cannot get an init id before saving the dataset, and we need an init id in order to save a workflow.

To ensure we do this correctly, besides just adding `wftest.TestCases`, that can be loaded from a directory using `LoadTestCasesFromDir` or from a default set of test cases using `LoadDefaultTestCases`, most test runners will probably want to use `MustAddDefaultWorkflows` or `MustAddWorkflowsFromDir`.

Both take a `wftest.TestRunner`, which just ensures we have all the pieces needed to align the workflow and dataset properly (such as correct peername & id), as well as save the dataset and workflow correctly.