refactor(collection): rely on `ETAutomationWorkflowCreated` events
We were previously using the `ETAutomationDeploy` events, but unfortunately they don't actually have enough specificity for us to know exactly what is going on when we catch the events in the collection.

We need to follow the same patterns we are using to add to the collection when a dataset is added or removed: rely on events emitted at those exact times.

We need an `ETAutomationWorkflowCreated` and an `ETAutomationWorkflowRemoved` event to let us know that a workflow has been added or removed.