refactor(transform): Single dataset, move conflict check up to lib.Save
Instead of a pair of read/write datasets on the starlark ds object, have only
a single dataset. For checking conflicts, have Apply track which components
are modified by a transform script. Compare that result against the components
provided to the save path, but do so in a much more obvious way, up in lib/