feat(collection): expand `WritableSet` interface, add option to pass in custom `WritableSet`
Because we need to repeat many of these handler/event patterns in `cloud`, we are dividing up the `Collection` to make it more reusable.

The `Collection` is now a struct that contains a `WritableSet`. The `Collection` is responsible for listening to events and calling various `Put`-like functions on the `set` when certain events are handled.

To keep things moving, rather than refactor a ton about the `collection` package, first I'm keeping these `Put`-like functions as close as possible to the methods that already existed on `localSet`. These can most likely be narrowed in future prs.

The `spec` package as been remove, and the tests moved into the `collection` packages itself. Although, `AssertWritableCollectionSpec` with some tweaking and additional tests should probably be moved back into a `spec` package.