feat(transform): `transform.Apply` & `transform.Commit`
We now have two consts that define all the current "modes" that
a transform can be run in: `RMApply` & `RMCommit`

Using the run mode we can distinguish between an "apply" (a
transform run with no intension of saving the result) and a "commit"
(a transform run with the intension of saving the result)

We can use the `transform.Apply` and `transform.Commit` methods to
indicate that we intend to apply or commit a transform. These
currently do the same actions (run a transform and return the dataset
preview). In the future, the `Commit` method may be modified to
do the actual saving of the output dataset.