refactor(event, transform): add `Mode` field to transform events
Added `Mode` to `TransformLifecycle`, `TransformStepLifecycle`, and `TransformMessage`

Event Types that have a `Mode` field in the payload:
	ETTransformStart = Type("tf:Start")
	ETTransformStop = Type("tf:Stop")
	ETTransformStepStart = Type("tf:StepStart")
	ETTransformStepStop = Type("tf:StepStop")
	ETTransformStepSkip = Type("tf:StepSkip")
	ETTransformError = Type("tf:Error")

Event Type that do NOT have a `Mode` field in the payload:
	ETTransformPrint = Type("tf:Print")
This event is only called via the `print` starlark command, not sure we need to be adding a "mode" to this, as I believe it's handled differently than other events in the frontend.

	ETTransformDatasetPreview = Type("tf:DatasetPreview")
This event has an event of type `dataset.Dataset`, and so we would need to add a new transform event to incorporate a `Mode` field into this event payload. However, I believe that it, too, is treated differently than other events in the frontend, and so we may not need to add a "mode" to this as well.