refactor(*): use transform events to track the number of runs
The `collection` can now track the number of "runs" that occur by using the `Mode` field in a transform event.

I'm choosing to use the `transform.ETTransformStart` event.

In order for the `collection` to correctly add the `RunCount` to a `VersionInfo`, it needs to know the `InitID` of the dataset. To that effect, we've also added an `InitID` field to the `event.TransformLifecycle` struct. If this field is empty, then the `collection` will log a debug error.