refactor(startf): transforms should return a dataset with the `Structure.Entries`
Users expect to see the row count in a dry run/apply. We can get the number of rows, aka the `dataset.Structure.Entries` count from the dataframe before we return the dataset.

However, the dataframe will only exist if the transform sets the body. If the transform does not set the body, we must instead load the previous structure and get the entries that way.

If there is no previous version, no previous structure, or no way to load the dataset, we do not add the previous `Structure.Entries` to the current dataset's `Structure`.