fix(transform): transforms default to CSV format
with our latest switch to using `ds.body` assignment, we've lost the `parse_as` argument that we formerly relied on to set data format. As a short-term fix we'd like to _default_ to CSV data format.

In a follow-up we'll detect the number of dimensions in the body and error if it's higher than 2.

closes #1915