feat(websocket): scope websocket connections
Websocket connections need to be scoped so we only send the relevant events to the correct connections.

To accomplish this, we keep track of the profile.ID of each connections using the new `websocket.conn` struct, as well as keeping a map of profile.IDs to connections.

To allow us to authenticate tokens, the `websocket.Handler` must have access to the `key.Store`.

We've added an authentication handshake. Once the connection has been established, the client can request to "subscribe" to a connection, and send over a token as part of the message payload. If that token is valid, we upgrade the connection, and send over a success message. If not, we send over a failure message.

If the client wants to unsubscribe, they send over an "unsubscribe" request. This will remove the association between that connection and a profile.ID. It will not close the connection.

We've also defined a `websocket.message` struct that the websocket expects to receive for communication about the state of the websocket connection (namely the authentication handshake).