refactor(automation): align contexts and run sync
First, we changed which methods are responsible for ensuring that we have a long-running context for the "run".

Then, we adjusted how the `ApplyWorkflow` handled the `wait` param. Since we need the `runLock` to be locked during the entire run, we need more fine grained control over when the `ApplyWorkflow` method returns. We now have a `applyWorkflow` method, which is run in a go routine when `wait == false`, and is run directly when otherwise.

This now means we can trust that functions run when we `defer` in `o.applyWorkflow` will trigger after the entire run of the transform.