feat(transform): Command `apply` sends terminal size to transform
The path from cmd / api down to dataframe is a bit tricky, but this PR pipes the size of the user's terminal down the stack as best it can. There are a few places where this isn't happening yet, in particular terminal info will not arrive for the `run` command, but that can happen in a follow-up PR. Users of the http api can send the size of their rendering target, such as the size of the html element, to affect dataframe rendering.