feat(whatchanged): Restore the `whatchanged` command
The command `whatchanged` returns what components changed at a specific version of in the history of the dataset.

This command was originally part of FSI, which was removed by commit eaeba31aeb6242bd4aacf79f7c20f9579585d549. However, this command is kind of its own thing, and does not have a hard dependency on FSI functionality. This change adds the command only, without the similar `status` command, which *did* depend on FSI.